unfpa-partners/requirements.txt
ipykernel
python-dotenv
pandas
openai
pydantic
requests

#pandas==2.2.3
playwright==1.51.0
Pillow==11.2.1
#openai==1.70.0
#pydantic==2.11.1
tabulate==0.9.0
#python-dotenv==1.1.0
#multion==1.3.8
#seaborn==0.13.2
#gradio==5.23.2
#pickleshare==0.7.5
#langsmith==0.3.21
#instructor==1.7.8
#litellm==1.65.1
#igraph==0.11.8
#plotly==6.0.1